fse-828: 결품현황 get method 추가

+++ src/main/kotlin/com/smartfoodnet/fninventory/stock/model/BasicProductStockModel.kt
@@ -15,7 +15,7 @@ data class BasicProductStockModel(
     val basicProductCode: String? = null,
 
     @ApiModelProperty(value = "출고상품 ID")
-    val shippingProductId: Int? = null,
+    val shippingProductId: Long? = null,
 
     @ApiModelProperty(value = "상품 바코드")
     val barcode: String? = null,
